Split redirect tests into two files.
